examples

+++ philosophers/philo/includes/philo.h
@@ -6,7 +6,7 @@
 /*   By: seungsle <seungsle@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/21 20:07:19 by seungsle          #+#    #+#             */
-/*   Updated: 2022/04/05 20:08:51 by seungsle         ###   ########.fr       */
+/*   Updated: 2022/04/05 21:16:30 by seungsle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,11 +21,11 @@
 # include <sys/time.h>
 
 typedef struct s_info{
-	int	number_of_philosophers;
-	int	time_to_die;
-	int	time_to_eat;
-	int	time_to_sleep;
-	int	number_of_times_each_philosopher_must_eat;
+	long long	number_of_philosophers;
+	long long	time_to_die;
+	long long	time_to_eat;
+	long long	time_to_sleep;
+	long long	number_of_times_each_philosopher_must_eat;
 }	t_info;
 
 size_t		ft_strlen(const char *s);
